backend/requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.9.1
blessed==1.22.0
channels==4.3.2
Django==4.2.24
django-allauth==0.61.1
django-cors-headers==4.9.0
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.1
dj-database-url==2.1.0
et_xmlfile==2.0.0
gunicorn==21.2.0
jinxed==1.3.0
openpyxl==3.1.5
pillow==12.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250822
tzdata==2025.2
wcwidth==0.2.13
whitenoise==6.6.0
